PlugAlgo: move replacePlug()
We move it into its own PlugAlgo so that we can reuse it.
