+++ src/GafferArnold/ParameterHandler.cpp
@@ -45,6 +45,7 @@
 #include "Gaffer/StringPlug.h"
 #include "Gaffer/Node.h"
 #include "Gaffer/ScriptNode.h"
+#include "Gaffer/PlugAlgo.h"
 
 #include "GafferArnold/ParameterHandler.h"
 
@@ -57,93 +58,6 @@ using namespace GafferArnold;
 
 namespace
 {
-
-struct Connections
-{
-	Plug *plug;
-	PlugPtr input;
-	vector<PlugPtr> outputs;
-};
-
-typedef vector<Connections> ConnectionsVector;
-
-void replacePlugWalk( Plug *existingPlug, Plug *plug, ConnectionsVector &connections )
-{
-	// Record output connections.
-	Connections c;
-	c.plug = plug;
-	c.outputs.insert( c.outputs.begin(), existingPlug->outputs().begin(), existingPlug->outputs().end() );
-
-	if( plug->children().size() )
-	{
-		// Recurse
-		for( PlugIterator it( plug ); !it.done(); ++it )
-		{
-			if( Plug *existingChildPlug = existingPlug->getChild<Plug>( (*it)->getName() ) )
-			{
-				replacePlugWalk( existingChildPlug, it->get(), connections );
-			}
-		}
-	}
-	else
-	{
-		// At a leaf - record input connection and transfer values if
-		// necessary. We only store inputs for leaves because automatic
-		// connection tracking will take care of connecting the parent
-		// levels when all children are connected.
-		c.input = existingPlug->getInput<Plug>();
-		if( !c.input && plug->direction() == Plug::In )
-		{
-			ValuePlug *existingValuePlug = runTimeCast<ValuePlug>( existingPlug );
-			ValuePlug *valuePlug = runTimeCast<ValuePlug>( plug );
-			if( existingValuePlug && valuePlug )
-			{
-				valuePlug->setFrom( existingValuePlug );
-			}
-		}
-	}
-
-	connections.push_back( c );
-}
-
-/// \todo Move to PlugAlgo.h and use in `OSLShader::loadShader()` too?
-void replacePlug( Gaffer::GraphComponent *parent, PlugPtr plug )
-{
-	Plug *existingPlug = parent->getChild<Plug>( plug->getName() );
-	if( !existingPlug )
-	{
-		parent->addChild( plug );
-		return;
-	}
-
-	// Transfer values where necessary, and store connections
-	// to transfer after reparenting.
-
-	ConnectionsVector connections;
-	replacePlugWalk( existingPlug, plug.get(), connections );
-
-	// Replace old plug by parenting in new one.
-
-	parent->setChild( plug->getName(), plug );
-
-	// Transfer old connections. We do this after
-	// parenting because downstream acceptsInput() methods
-	// might care what sort of node the connection is coming
-	// from.
-
-	for( ConnectionsVector::const_iterator it = connections.begin(), eIt = connections.end(); it != eIt; ++it )
-	{
-		if( it->input )
-		{
-			it->plug->setInput( it->input.get() );
-		}
-		for( vector<PlugPtr>::const_iterator oIt = it->outputs.begin(), oeIt = it->outputs.end(); oIt != oeIt; ++oIt )
-		{
-			(*oIt)->setInput( it->plug );
-		}
-	}
-
-}
 
 template<typename PlugType>
 Gaffer::Plug *setupNumericPlug( const AtNodeEntry *node, const AtParamEntry *parameter, Gaffer::GraphComponent *plugParent, Gaffer::Plug::Direction direction )
@@ -187,7 +101,7 @@ Gaffer::Plug *setupNumericPlug( const AtNodeEntry *node, const AtParamEntry *par
 	}
 
 	typename PlugType::Ptr plug = new PlugType( name, direction, defaultValue, minValue, maxValue, Plug::Default | Plug::Dynamic );
-	replacePlug( plugParent, plug );
+	PlugAlgo::replacePlug( plugParent, plug );
 
 	return plug.get();
 }
@@ -205,7 +119,7 @@ Gaffer::Plug *setupPlug( const IECore::InternedString &parameterName, Gaffer::Gr
 	}
 
 	PlugPtr plug = new Plug( parameterName, direction, Plug::Default | Plug::Dynamic );
-	replacePlug( plugParent, plug );
+	PlugAlgo::replacePlug( plugParent, plug );
 
 	return plug.get();
 }
@@ -226,7 +140,7 @@ Gaffer::Plug *setupTypedPlug( const IECore::InternedString &parameterName, Gaffe
 	typename PlugType::Ptr plug = new PlugType( parameterName, direction, defaultValue );
 	plug->setFlags( Plug::Dynamic, true );
 
-	replacePlug( plugParent, plug );
+	PlugAlgo::replacePlug( plugParent, plug );
 
 	return plug.get();
 }
@@ -278,7 +192,7 @@ Gaffer::Plug *setupColorPlug( const AtNodeEntry *node, const AtParamEntry *param
 	}
 
 	typename PlugType::Ptr plug = new PlugType( name, direction, defaultValue, minValue, maxValue, Plug::Default | Plug::Dynamic );
-	replacePlug( plugParent, plug );
+	PlugAlgo::replacePlug( plugParent, plug );
 
 	return plug.get();
 }
